systemd-compat-units: do not create runlevel command
It seems strange that runlevel always returns 1, The comment says it is
related to sysvinit, but if we enable systemd, sysvinit will not be
installed. and we have created a link for runlevel to systemctl if
systemd is installed.

(From OE-Core rev: adc11d60fd4b555198d6653cd71eb1372e0b03a0)

Signed-off-by: Roy.Li <rongqing.li@windriver.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>