systemd: add packageconfig for sysext
Currently the sysext binary is packaged in systemd-utils and
the service systemd-sysext.service in systemd package. Add a
PACKAGECONFIG for sysext to actively choose if it is going to
be installed or not, default off. If installed it will be added
to systemd package.

(From OE-Core rev: 98cc5639d74dc71aa1352b50f97ed832245d3c3b)

Signed-off-by: Peter Bergin <peter@berginkonsult.se>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>